removed lifetime from GuestMemory
This is one way we can avoid having to specify a lifetime parameter for
the GuestMemory trait. The parameter is now associated with the
concrete type which implements GuestMemory, but this can be made
relatively painless by passing values or references to
`impl GuestMemory`, as opposed to referring to the actual type, with
its newly acquired lifetime parameter.

Signed-off-by: Alexandru Agache <aagch@amazon.com>